backend-python/pyproject.toml
[project]
name = "prompt-stress-tester"
version = "1.0.0"
description = "AI Prompt Stress Tester Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "boto3>=1.35.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
